Proposed new section and other improvements
I think the only potentially controversial change I proposed is the removal of "Don't quit your job over this."  I think some people will want to quit their jobs if they realize their employers do not care about their health or the health of their customers.  I've heard many complaints about people who have been very upset about being "forced" or "pressured" to do things that are not the best for them or society, and are only best for profits.  For most people (and for global health), it's likely better to be unemployed and alive, than employed and dead or infecting others.